requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.3
bitcoin==1.1.42
bitcoinlib==0.6.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cytoolz==0.12.1
ecdsa==0.18.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
hexbytes==0.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
parsimonious==0.9.0
pip==22.3.1
protobuf==4.22.3
psycopg2==2.9.6
pycryptodome==3.17
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
setuptools==65.5.1
six==1.16.0
SQLAlchemy==2.0.9
toolz==0.12.0
typing-extensions==4.5.0
urllib3==1.26.15
web3==6.2.0
websockets==11.0.2
Werkzeug==2.2.3
wheel==0.38.4
yarl==1.8.2
